+++ src/Vector3.java
@@ -1,14 +1,14 @@
 public class Vector3{
 	//TODO convert to use ICoord, Number is too slow.
 	private static byte dim = 3;
-	private int[] vals;
+	private float[] vals;
 	
-	public Vector3(int[] vals) {
+	public Vector3(float[] vals) {
 		this.vals = vals;
 	}
 	
 	public Vector3 subtract(Vector3 second) {
-		int[] data = new int[3];
+		float[] data = new float[3];
 		for (int i = 0; i < dim; i++) {
 			data[i] = vals[i] - second.vals[i];
 		}
@@ -17,18 +17,18 @@ public class Vector3{
 	
 	public Vector3 add(Vector3 second) {
 		
-		int[] data = new int[3];
+		float[] data = new float[3];
 		for (int i = 0; i < dim; i++) {
 			data[i] = vals[i] + second.vals[i];
 		}
 		return new Vector3(data);
 	}
 	
-	public int getValue(byte i) {
+	public float getValue(int i) {
 		return vals[i];
 	}
 
-	public void setValue(byte i, int value) {
+	public void setValue(int i, float value) {
 		vals[i] = value;
 	}
 }
